requirements.txt
langchain>=0.1.16
langchain-community>=0.0.36
langchain-core>=0.1.48
langchain-groq>=0.1.4
groq>=0.3.1
langchain-openai>=0.1.0
openai>=1.0.0
wikipedia>=1.4.0
arxiv>=2.1.0
duckduckgo-search>=5.3.0
streamlit>=1.31.1
pydantic>=2.6.4
python-dotenv>=1.0.1
tiktoken>=0.6.0
